Improve implementation of NSURLRequest and NSMutableURLRequest.
By default, network access is disabled and we return nil for init of
NSURLRequest, same as before. This should prevent regressions in the
currently supported apps, but it could be removed once we support more
methods of NSURLRequest, as constructing request itself have nothing to
do with network access by itself.

Change-Id: I943e35806db3884f4dc4479617b764030061bf22