Remove dylib symlinks from git
touchHLE doesn't actually use these (the virtual filesystem's
dylib symlinks are manually enumerated in fs.rs), but they do cause
problems on Windows (which generally doesn't like symlinks), and cause
release bloat because they get turned into unused duplicate files.

The only real use for them remaining is when analysing stuff in Ghidra
or the like. Maybe we could add some script to re-create the symlinks
for that purpose, but I've not done that here.

This commit also fixes the dylibs README not mentioning libz aliases.

Change-Id: I79aadc0620dbc7e4e72d9b08629490ad2d72af51