Use DocumentsProvider.getContext() to get context
This must have been an oversight when implementing the documents
provider (6d714bbf0d2c64743c7b2b24e9c5af479b17631c). There is a
method on DocumentsProvider that gives us access to the context, so
trying to get it via the Application is unnecessary, and we can delete
that code. Apparently this also fixes the mysterious crashes when
opening the file manager the first few times.

Change-Id: Iab674e05104a7c895a2df2a2fcf9d0f1942f125b