Reorganise host class, constant and function exports into "dylibs"
This replaces the three huge lists of lists for host classes, constants
and functions with a single list of host "dylibs". Each dylib has its
own lists of lists for classes, constants, and functions, plus
information about the name/path(s) the dylib is referenced by.

The key benefit to this change is that now touchHLE knows which library
each symbol belongs to. There are several different things we might want
to use this for in future, but this commit doesn't implement any of them;
it is a "no functional change" refactoring for now.

This commit also adds a test to detect duplicate constant definitions,
mirroring the existing tests for duplicate classes and functions.

Change-Id: I75f3c32aeb418a0162c97f3a628ce48f600b6ad2