Support point/rect conversion to/from screen-space co-ordinates
This fixes the incorrect offset applied to touch input handling for
windows that aren't positioned at (0,0), and adds support for:

- CALayer convertPoint:fromLayer: with nil layer parameter
- CALayer convertPoint:toLayer: with nil layer parameter
- UIWindow convertPoint:fromWindow:
- UIWindow convertPoint:toWindow:

The CALayer visual tests were modified to enable partial testing of
this. This was also tested by the fact the GUI tests part of TestApp
happened to be affected by the incorrect offset in touch handling.

Change-Id: I00aab7f0b63bc87c8b630e433fd87b06a8925c3f