Support CALayer affineTransform and UIView transform
Fixes #388. The lack of support for transforms has been a longstanding
limitation of the compositor and code for co-ordinate space conversion
and hit testing. This will fix many games that need these for rotating
individual UI elements, or the entire game screen, and will enable
eventually supporting iOS's support for automatic rotation of UI.

This commit also improves bounds.origin support, and adds support for
-[CALayer convertPoint:toLayer:] and -[CALayer convertPoint:fromLayer:].

Change-Id: I6b421e2f853d50c4742f851f61117c2b6545a3ba