Make windows hidden by default and keep track of invisible windows
This bug was noticed while adding the UIKit section to TestApp: on real
iOS, the window doesn't show until you explicitly make it visible, i.e.
it is hidden by default.

The change to keep track of invisible windows is a small improvement in
correctness that happened to simplify the main change here.

Change-Id: I8aa3386c7cc2ed3252512cb38ac7c71685a41c79